# app/api/v1/endpoints/settings.py
import logging
import traceback
from typing import Any

from fastapi import APIRouter, Depends, HTTPException
from sqlalchemy.exc import SQLAlchemyError
from sqlalchemy.orm import Session

from app import schemas
from app.api import deps
from app.services import create_settings, get_settings, update_settings

logger = logging.getLogger(__name__)
router = APIRouter(prefix="/settings")


@router.get("/", response_model=schemas.Settings)
def get_settings_endpoint(
    db: Session = Depends(deps.get_db),
    current_user: schemas.User = Depends(deps.get_current_active_user),
) -> Any:
    """
    Get current user settings
    """
    try:
        logger.info(f"Getting user settings: user_id={current_user.id}")
        settings = get_settings(db, user_id=current_user.id)
        if not settings:
            logger.info(f"User settings not found: user_id={current_user.id}")
            raise HTTPException(status_code=404, detail="Settings not found")
        logger.info(f"Successfully retrieved user settings: user_id={current_user.id}")
        return settings
    except SQLAlchemyError as e:
        logger.error(
            f"Failed to get user settings: user_id={current_user.id}, error={str(e)}"
        )
        logger.error(f"Error stack: {traceback.format_exc()}")
        raise HTTPException(status_code=500, detail="Database operation failed")
    except Exception as e:
        logger.error(
            f"Unknown error occurred while getting user settings: user_id={current_user.id}, error={str(e)}"
        )
        logger.error(f"Error stack: {traceback.format_exc()}")
        raise HTTPException(status_code=500, detail="Internal server error")


@router.put("/", response_model=schemas.Settings)
def update_settings_endpoint(
    *,
    db: Session = Depends(deps.get_db),
    settings_in: schemas.SettingsUpdate,
    current_user: schemas.User = Depends(deps.get_current_active_user),
) -> Any:
    """
    Update current user settings
    """
    try:
        logger.info(
            f"Updating user settings: user_id={current_user.id}, settings={settings_in.dict()}"
        )
        settings = get_settings(db, user_id=current_user.id)
        if not settings:
            # Create new settings
            settings_create = schemas.SettingsCreate(
                language=settings_in.language or "en",
                voice_enabled=(
                    settings_in.voice_enabled
                    if settings_in.voice_enabled is not None
                    else True
                ),
            )
            logger.info(
                f"Creating new user settings: user_id={current_user.id}, settings={settings_create.dict()}"
            )
            settings = create_settings(
                db, settings_in=settings_create, user_id=current_user.id
            )
        else:
            logger.info(
                f"Updating existing user settings: user_id={current_user.id}, settings={settings_in.dict()}"
            )
            settings = update_settings(
                db, db_settings=settings, settings_in=settings_in
            )
        logger.info(f"Successfully updated user settings: user_id={current_user.id}")
        return settings
    except SQLAlchemyError as e:
        logger.error(
            f"Failed to update user settings: user_id={current_user.id}, error={str(e)}"
        )
        logger.error(f"Error stack: {traceback.format_exc()}")
        raise HTTPException(status_code=500, detail="Database operation failed")
    except Exception as e:
        logger.error(
            f"Unknown error occurred while updating user settings: user_id={current_user.id}, error={str(e)}"
        )
        logger.error(f"Error stack: {traceback.format_exc()}")
        raise HTTPException(status_code=500, detail="Internal server error")