# app/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "InTy",
        "description": "InTy",
        "version": "1.0.2"
    },
    "paths": {
...
        "/api/v1/report/upload-image": {
            "post": {
                "tags": [
                    "report"
                ],
                "summary": "Upload image for report",
                "description": "Used by app to upload image in their report of app content: AI characters, images, etc.",
                "operationId": "upload_report_image_api_v1_report_upload_image_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_report_image_api_v1_report_upload_image_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse_dict_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/report/": {
            "post": {
                "tags": [
                    "report"
                ],
                "summary": "Create Report",
                "description": "Submit report",
                "operationId": "create_report_api_v1_report__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReportCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/api/v1/settings/": {
            "get": {
                "tags": [
                    "settings"
                ],
                "summary": "Get Settings Endpoint",
                "description": "Get current user settings",
                "operationId": "get_settings_endpoint_api_v1_settings__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Settings"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "settings"
                ],
                "summary": "Update Settings Endpoint",
                "description": "Update current user settings",
                "operationId": "update_settings_endpoint_api_v1_settings__put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SettingsUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Settings"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "APIResponse": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "success"
                    },
                    "data": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data"
                    }
                },
                "type": "object",
                "title": "APIResponse"
            },
...
            "APIResponse_dict_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "success"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data"
                    }
                },
                "type": "object",
                "title": "APIResponse[dict]"
            },
...
            "Body_upload_report_image_api_v1_report_upload_image_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_report_image_api_v1_report_upload_image_post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ReportCreate": {
                "properties": {
                    "target_id": {
                        "type": "string",
                        "title": "Target Id"
                    },
                    "target_type": {
                        "$ref": "#/components/schemas/TargetType"
                    },
                    "reason_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "type": "array",
                        "title": "Reason Ids"
                    },
                    "image_urls": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Urls",
                        "default": []
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    }
                },
                "type": "object",
                "required": [
                    "target_id",
                    "target_type",
                    "reason_ids"
                ],
                "title": "ReportCreate"
            },
            "Settings": {
                "properties": {
                    "language": {
                        "type": "string",
                        "title": "Language",
                        "default": "en"
                    },
                    "voice_enabled": {
                        "type": "boolean",
                        "title": "Voice Enabled",
                        "default": true
                    },
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "user_id": {
                        "type": "string",
                        "title": "User Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "user_id",
                    "created_at"
                ],
                "title": "Settings",
                "description": "API 响应中的设置"
            },
            "SettingsUpdate": {
                "properties": {
                    "language": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Language"
                    },
                    "voice_enabled": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Voice Enabled"
                    }
                },
                "type": "object",
                "title": "SettingsUpdate",
                "description": "更新设置"
            },
...
            "TargetType": {
                "type": "string",
                "enum": [
                    "USER",
                    "AGENT"
                ],
                "title": "TargetType"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "\n            输入格式为: your_token\n            注意: 不需要Bearer前缀\n            "
            }
        }
    }
}